feat: showNotification 함수를 외부로 분리
